+++ sistema.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+#! sistema.py
 
 from datetime import datetime, timedelta
 from decimal import Decimal
@@ -8,19 +9,8 @@ from sys import argv
 import pandas as pd
 import pandas_ta as ta
 
-workerTypes = ["test","dbWorker", "dbMiner", "dbCalculator", "MACDentry"]
-
+workerTypes = ["dbWorker", "dbMiner", "dbCalculator"]
 db = DB()
-
-def test_general():
-	print("|TESTING CONTAINER ENVIRONMENT")
-
-def test():
-	a = Worker("test", "test")
-	a.updateTime = timedelta(minutes=1)
-	while True:
-		a._internalTick()
-
 class Worker:
 	def __init__(self, user, workType):
 		self.API = db.getAPI(user)
@@ -134,48 +124,7 @@ class dbCalculator(Worker):
 					self._calculate(pair["symbol"], "1d")
 				self.lastCheck = db.getOlderServe(self.work)
 
-class MACDentry(Worker):
-	def __init__(self, user, workType):
-		super().__init__(user, workType)
-		self.updateTime = timedelta(seconds=15)
-		self.maxOld = timedelta(hours=4)
-	def _checkDate(self, df):
-		"""Comprueba que la fecha no es más lejana de self.maxOld
-
-		Args:
-			df ([type]): [description]
-		"""
-		pass
-	def startWork(self):
-		self.lastCheck = db.getOlderServe(self.work)
-		while True:
-			if self._internalTick() == True:
-				pairs = db.servePairs(self.work, limit=100)
-				for pair in pairs:
-					df4h = db.getDataFrame(pair["symbol"], "4h")
-					if df4h.empty == False:
-						last4h = df4h["histogram"].iat[-1]
-						prelast4h = df4h["histogram"].iat[-2]
-						if last4h is not None and prelast4h is not None: 
-							if last4h > prelast4h:
-								if last4h > 0:
-									#print(Decimal(df4h["histogram"].iat[-1]))
-									price = Decimal(self.client.get_symbol_ticker(symbol=pair["symbol"])["price"])
-									print(f'{pair["symbol"]}: {price}')
-									print(df4h["openTime"].iat[-1])
-									print("---> Abriendo Trade!")
-								else:
-									pass
-							else:
-								pass
-						else:
-							print("Cant Check histogram, NoneValue")
-					else:
-						print("Dataframe empty")
-				self.lastCheck = db.getOlderServe(self.work)
-
 if __name__ == "__main__":
-	#test_general()
 	##argv1 = USER/test
 	##argv2 = workerType/testType
 	print(datetime.now())
@@ -190,8 +139,6 @@ if __name__ == "__main__":
 					worker = dbMiner(argv[1], argv[2])
 				elif argv[2] == "dbCalculator":
 					worker = dbCalculator(argv[1], argv[2])
-				elif argv[2] == "MACDentry":
-					worker = MACDentry(argv[1], argv[2])
 				try:
 					worker.startWork()
 				except KeyboardInterrupt:
