+++ staging.yml
@@ -19,5 +19,23 @@ services:
       - 8000:8080
     depends_on: 
       - mariadb
+  dbworker:
+    build:
+      dockerfile: worker.dockerfile
+    command: sistema.py david dbWorker
+    depends_on:
+      - mariadb
+  dbminer:
+    build:
+      dockerfile: worker.dockerfile
+    command: sistema.py david dbMiner
+    depends_on:
+      - mariadb
+  dbcalculator:
+    build: 
+      dockerfile: worker.dockerfile
+    command: sistema.py david dbCalculator
+    depends_on: 
+      - mariadb
 volumes:
   dbDataDEV:
